fix: change AIDecision.Confidence from int to float64
Position Manager returns confidence as float (0.6), which failed
to unmarshal into int field. Now accepts both int and float values.

+++ internal/ai/types.go
@@ -92,7 +92,7 @@ type AIDecision struct {
 	Ticker     string  `json:"ticker"`
 	StopLoss   float64 `json:"stop_loss"`
 	TakeProfit float64 `json:"take_profit"`
-	Confidence int     `json:"confidence"` // 0-100
+	Confidence float64 `json:"confidence"` // 0-100 (screening) or 0-1 (position manager)
 	Reasoning  string  `json:"reasoning"`
 }
